requirements.txt
opencv-python==4.5.5.64
face_recognition==1.3.0
Flask==2.0.2
PyYAML==6.0
Pillow==9.0.1
redis==3.5.3
flask-debug==0.4.3
watchdog==2.1.6